- 회의 참가, 종료 시각, 회의 시간 기록 기능 추가  - 참가자 REDIS 데이터 제거  - 사용자별 TALK 시간 기록 기능 추가

+++ src/main/java/com/mindplates/everyonesprint/biz/meeting/vo/response/MeetingSummaryResponse.java
@@ -19,7 +19,7 @@ public class MeetingSummaryResponse {
     private LocalDateTime endDate;
     private LocalDateTime realStartDate;
     private LocalDateTime realEndDate;
-    private Integer durationSeconds;
+    private Long durationSeconds;
     private List<User> users;
     private Long sprintDailyMeetingId;
 
@@ -50,7 +50,7 @@ public class MeetingSummaryResponse {
         private Long userId;
         private LocalDateTime firstJoinDate;
         private LocalDateTime lastOutDate;
-        private Integer joinDurationSeconds;
-        private Integer talkedSeconds;
+        private Long joinDurationSeconds;
+        private Long talkedSeconds;
     }
 }
